pyproject.toml
[project]
name = "satellite-image-analysis"
version = "2.3.0"
description = "Satellite land cover analysis - analyze('Florence') does everything"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Vito Delia", email = "vito.delia@example.com" }
]
keywords = ["satellite", "sentinel", "remote-sensing", "machine-learning", "clustering", "land-cover", "pipeline"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "numpy>=1.24.0",
    "rasterio>=1.3.0",
    "matplotlib>=3.7.0",
    "pillow>=10.0.0",
    "scikit-learn>=1.3.0",
    "scipy>=1.8.0",
    "geopy>=2.3.0",
    "shapely>=2.0.0",
    "pyproj>=3.5.0",
    "utm>=0.7.0",
    "pyyaml>=6.0.0",
    "requests>=2.31.0",
    "requests-oauthlib>=1.3.1",
    "oauthlib>=3.2.2",
    "tqdm>=4.66.0",
]

[project.optional-dependencies]
dev = [
    "jupyter>=1.0.0",
    "ipykernel>=6.25.0",
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]

ui = [
    "streamlit>=1.28.0",
]

api = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "python-multipart>=0.0.6",
]

notebooks = [
    "jupyter>=1.0.0",
    "ipykernel>=6.25.0",
    "nbformat>=5.9.0",
]

annotation = [
    "labelme>=5.3.0",
    "pyside6>=6.5.0",
]

all = [
    "satellite-image-analysis[dev,ui,api,notebooks,annotation]",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/satellite_analysis"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests requiring real data",
    "network: marks tests as requiring network access",
]
